+++ CHANGELOG.md
@@ -9,10 +9,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Added a warning message when the customer schema is invalid
+- Added a warning message when the customer schema is invalid.
+
+## [1.11.0] - 2022-06-23
+
+## [1.10.3] - 2022-06-22
 
 ### Added
 
+- Add error messages improvement
+
+### Added
+
+- UI support for organization `tradeName` and cost center `phoneNumber` fields
 - Github Action to trigger manual tests by dispatch
 
 ## [1.10.2] - 2022-06-14